// src/app/models/modelsInterfaces.ts
export class PokemonDetailed {
  abilities: any[] = [];
  id: number = 0;
  cries: any[] = [];
  imgUrl: string = '';
  types: any[] = [];
  name: string = '';
}